refactor(cli): move MCP Apps logic into Kilo-owned module
Extract the experimental MCP Apps schemas, HTTP handler bodies, and the
tool-call metadata helper into @/kilocode/mcp/apps so the shared upstream
files only carry minimal marked integration points (route registrations,
imports, one field). Reduces the long-term diff against upstream opencode.
